+++ twint-action/src/test/java/com/adyen/checkout/twint/action/TwintActionConfigurationTest.kt
@@ -1,0 +1,90 @@
+/*
+ * Copyright (c) 2024 Adyen N.V.
+ *
+ * This file is open source and available under the MIT license. See the LICENSE file for more info.
+ *
+ * Created by oscars on 9/7/2024.
+ */
+
+package com.adyen.checkout.twint.action
+
+import com.adyen.checkout.components.core.Amount
+import com.adyen.checkout.components.core.AnalyticsConfiguration
+import com.adyen.checkout.components.core.AnalyticsLevel
+import com.adyen.checkout.components.core.CheckoutConfiguration
+import com.adyen.checkout.core.Environment
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
+import java.util.Locale
+
+internal class TwintActionConfigurationTest {
+
+    @Test
+    fun `when creating the configuration through CheckoutConfiguration, then it should be the same as when the builder is used`() {
+        val checkoutConfiguration = CheckoutConfiguration(
+            shopperLocale = Locale.US,
+            environment = Environment.TEST,
+            clientKey = TEST_CLIENT_KEY,
+            amount = Amount("EUR", 123L),
+            analyticsConfiguration = AnalyticsConfiguration(AnalyticsLevel.ALL),
+        ) {
+            twintAction()
+        }
+
+        val actual = checkoutConfiguration.getTwintActionConfiguration()
+
+        val expected = TwintActionConfiguration.Builder(
+            shopperLocale = Locale.US,
+            environment = Environment.TEST,
+            clientKey = TEST_CLIENT_KEY,
+        )
+            .setAmount(Amount("EUR", 123L))
+            .setAnalyticsConfiguration(AnalyticsConfiguration(AnalyticsLevel.ALL))
+            .build()
+
+        assertEquals(expected.shopperLocale, actual?.shopperLocale)
+        assertEquals(expected.environment, actual?.environment)
+        assertEquals(expected.clientKey, actual?.clientKey)
+        assertEquals(expected.amount, actual?.amount)
+        assertEquals(expected.analyticsConfiguration, actual?.analyticsConfiguration)
+    }
+
+    @Test
+    fun `when the configuration is mapped to CheckoutConfiguration, then CheckoutConfiguration is created correctly`() {
+        val config = TwintActionConfiguration.Builder(
+            shopperLocale = Locale.US,
+            environment = Environment.TEST,
+            clientKey = TEST_CLIENT_KEY,
+        )
+            .setAmount(Amount("EUR", 123L))
+            .setAnalyticsConfiguration(AnalyticsConfiguration(AnalyticsLevel.ALL))
+            .build()
+
+        val actual = config.toCheckoutConfiguration()
+
+        val expected = CheckoutConfiguration(
+            shopperLocale = Locale.US,
+            environment = Environment.TEST,
+            clientKey = TEST_CLIENT_KEY,
+            amount = Amount("EUR", 123L),
+            analyticsConfiguration = AnalyticsConfiguration(AnalyticsLevel.ALL),
+        )
+
+        assertEquals(expected.shopperLocale, actual.shopperLocale)
+        assertEquals(expected.environment, actual.environment)
+        assertEquals(expected.clientKey, actual.clientKey)
+        assertEquals(expected.amount, actual.amount)
+        assertEquals(expected.analyticsConfiguration, actual.analyticsConfiguration)
+
+        val actualAwaitConfig = actual.getTwintActionConfiguration()
+        assertEquals(config.shopperLocale, actualAwaitConfig?.shopperLocale)
+        assertEquals(config.environment, actualAwaitConfig?.environment)
+        assertEquals(config.clientKey, actualAwaitConfig?.clientKey)
+        assertEquals(config.amount, actualAwaitConfig?.amount)
+        assertEquals(config.analyticsConfiguration, actualAwaitConfig?.analyticsConfiguration)
+    }
+
+    companion object {
+        private const val TEST_CLIENT_KEY = "test_qwertyuiopasdfghjklzxcvbnmqwerty"
+    }
+}